[MTDSA-922] Refactor auth code.
