+++ app/uk/gov/hmrc/selfassessmentapi/resources/BaseResource.scala
@@ -18,63 +18,34 @@ package uk.gov.hmrc.selfassessmentapi.resources
 
 import play.api.Logger
 import play.api.mvc.{RequestHeader, Result}
-import uk.gov.hmrc.auth.core._
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.play.http.HeaderCarrier
 import uk.gov.hmrc.play.microservice.controller.BaseController
-import uk.gov.hmrc.selfassessmentapi.config.MicroserviceAuthConnector
 import uk.gov.hmrc.selfassessmentapi.connectors.BusinessDetailsConnector
-import uk.gov.hmrc.selfassessmentapi.controllers.definition.HttpMethod
-import uk.gov.hmrc.selfassessmentapi.models.Errors.Error
-import uk.gov.hmrc.selfassessmentapi.models.MtdId
+import uk.gov.hmrc.selfassessmentapi.services.AuthenticatorService
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-trait BaseResource extends BaseController with AuthorisedFunctions {
-  override def authConnector: AuthConnector = MicroserviceAuthConnector
+trait BaseResource extends BaseController {
   private val businessConnector = BusinessDetailsConnector
+  private val authService = AuthenticatorService
 
-  val logger: Logger
+  val logger: Logger = Logger(this.getClass)
 
   def withAuth(nino: Nino)(f: => Future[Result])
               (implicit hc: HeaderCarrier, reqHeader: RequestHeader): Future[Result] =
     businessConnector.get(nino).flatMap { response =>
       response.status match {
-        case 200 => authorise(response.mtdId)(f)
-        case 400 | 404 => Future.successful(Unauthorized)
+        case 200 => {
+          logger.debug(s"NINO to MTD reference lookup successful. Status Code ${response.status}.")
+          authService.authorise(response.mtdId)(f)
+        }
+        case 400 | 404 => {
+          logger.debug(s"NINO to MTD reference lookup failed. Status Code ${response.status}")
+          Future.successful(Unauthorized)
+        }
         case _ => Future.successful(unhandledResponse(response.status, logger))
       }
     }
-
-  private def authorise(mtdId: Option[MtdId])(f: => Future[Result])
-                       (implicit hc: HeaderCarrier, reqHeader: RequestHeader): Future[Result] =
-    mtdId match {
-      case Some(id) => authoriseAsClient(id)(f)
-      case None => Future.successful(Unauthorized)
-    }
-
-  private def authoriseAsClient(mtdId: MtdId)(f: => Future[Result])
-                               (implicit hc: HeaderCarrier, requestHeader: RequestHeader): Future[Result] =
-    authorised(
-      Enrolment("HMRC-MTD-IT")
-        .withIdentifier("MTDITID", mtdId.mtdId)
-        .withDelegatedAuthRule("mtd-it-auth")) {
-      f
-    } recoverWith authoriseAsFOA(f) recoverWith unauthorised
-
-  private def authoriseAsFOA(f: => Future[Result])
-                            (implicit hc: HeaderCarrier, reqHeader: RequestHeader): PartialFunction[Throwable, Future[Result]] = {
-    case _: InsufficientEnrolments => authorised(
-      Enrolment("HMRC-AS-AGENT")) {
-      if (reqHeader.method == "GET") Future.successful(Unauthorized) else f
-    }
-  }
-
-  private val unauthorised: PartialFunction[Throwable, Future[Status]] = {
-    case e: AuthorisationException => {
-      logger.debug(s"Client authorisation failed. Exception: [$e]")
-      Future.successful(Unauthorized)
-    }
-  }
 }
