+++ func/uk/gov/hmrc/selfassessmentapi/featureswitch/SelfEmploymentFeatureSwitchSpec.scala
@@ -20,6 +20,7 @@ class SelfEmploymentFeatureSwitchSpec extends BaseFunctionalSpec {
   "self-employments" should {
     "not be visible if feature Switched Off" in {
       given()
+        .userIsSubscribedToMtdFor(nino)
         .userIsFullyAuthorisedForTheResource(nino)
         .when()
         .get(s"/ni/$nino/${SourceType.SelfEmployments.toString}")
